LayoutTests/http/tests/xmlhttprequest/access-control-basic-whitelist-request-headers-expected.txt

Accept: *
Accept-Language: ru
Content-Language: ru
Content-Type: text/plain; charset=UTF-8
